requirements.txt
streamlit>=1.31.0
pyyaml>=6.0.1
reportlab>=4.0.9
